packages/syrviscore/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "syrviscore"
version = "0.1.21"
description = "SyrvisCore - Self-hosted infrastructure platform for Synology NAS"
readme = "README.md"
authors = [
    { name = "Kevin Tegtmeier", email = "your-email@example.com" }
]
license = { text = "MIT" }
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "click==8.1.7",
    "pyyaml==6.0.2",
    "requests==2.31.0",
    "docker==7.1.0",
    "python-dotenv==1.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "build>=0.10.0",
]

[project.scripts]
syrvis = "syrviscore.cli:cli"

[project.urls]
Homepage = "https://github.com/kevinteg/SyrvisCore"
Documentation = "https://github.com/kevinteg/SyrvisCore/wiki"
Repository = "https://github.com/kevinteg/SyrvisCore"
Issues = "https://github.com/kevinteg/SyrvisCore/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
syrviscore = ["py.typed", "build/*.yaml"]

[tool.black]
line-length = 100
target-version = ["py38"]

[tool.ruff]
line-length = 100
target-version = "py38"
